thêm hủy phiếu đặt trước

+++ NTH_Restaurant_API/src/main/java/com/example/NTH_Restaurant_API/service/Impl/CT_BanServiceImpl.java
@@ -83,6 +83,9 @@ public class CT_BanServiceImpl implements CT_BanService {
 
     @Override
     public String xoaCT_Ban(Integer idCTB) {
+        if(!ct_banRepository.existsById(idCTB)){
+            return "false";
+        }
         try {
             ct_banRepository.deleteById(idCTB);
             return "true";
